Changing code organization
- Removing unnecessary imports
- Fixing a bug in TrainAndTestGradientQL.py where the variable being returned was wrong
- Splitting the neural network QL and the table QL in different modules so the table does not need PyTorch.
