+++ .gitignore
@@ -10,3 +10,4 @@ PrintOptimalPolicies.py
 OptimalPolicies.txt
 
 DebugScript.py
+ReadAndResave.py
